block: vhdx: Rename parser.rs to dynamic.rs
The file no longer holds a parser. It contains the Vhdx disk object with
its positioned read and write path, the Sector translation helper and
the VhdxError and VhdxIoError types. VHDX only implements the dynamic
subformat here, so name the file after it, matching how vhd keeps the
fixed disk object in fixed.rs.
