+++ block/src/formats/vhdx/mod.rs
@@ -10,10 +10,10 @@
 //! images.
 
 mod bat;
+mod dynamic;
 mod engine_sync;
 mod header;
 mod metadata;
-mod parser;
 #[cfg(test)]
 mod test_util;
 
@@ -22,7 +22,7 @@ use std::io::Error as IoError;
 use std::os::fd::AsRawFd;
 use std::sync::{Arc, Mutex};
 
-pub use parser::{Vhdx, VhdxError};
+pub use dynamic::{Vhdx, VhdxError};
 
 use self::engine_sync::VhdxSync;
 use crate::async_io::{AsyncIo, BorrowedDiskFd, DiskFileError};
